fix: Overrides when using hashtag (#84)
When we introduced elevated rate limits, we also implemented the use of Redis Hashtags to ensure that all keys involved in the elevated rate limits operation are allocated in the same Redis hash slot. This hashtagging process involves adding curly braces around the key.

However, this introduced an issue: overrides typically target keys without the hashtag. As a result, if we use a key with hashtags, the overrides will not be applied.

This PR addresses the issue by sanitizing the received key by removing the hashtag before looking for overrides. If any overrides are found, they will be applied accordingly.

Example:

Given the following configuration

buckets: {
  ip: {
    size: 10,
    per_second: 5,
    overrides: {
      '127.0.0.1': {
        per_second: 100
      }
    },
  }
}
It will apply the overrides for:

take('ip', '127.0.0.1')
take('ip', '{127.0.0.1}')
It will not apply the overrides for any other combination.