chore: removed ping in favor of ioredis keepalive (#86)
We observed that `limitd-redis` fails to promptly detect a stale TCP connection to Redis, only realizing the connection is down after approximately 12 minutes. After this delay, it reconnects as expected.

This update enables the `keepAlive` feature in ioredis, which sets the `KEEPALIVE` option on the socket. This allows for a quicker reaction to stale connections.

Key changes:
- The `keepAlive` feature is enabled by default and set to 10000 ms.
- The `keepAlive` interval can be customized by passing the `keepAlive` argument to the `LimitdRedis` constructor.
- Setting the `keepAlive` argument to a non-numeric value will disable the feature.
- The manually-implemented ping mechanism against Redis has been removed, as we now rely on ioredis' `keepAlive` feature.